+++ ChangeLog.md
@@ -7,6 +7,8 @@
 ### Enhancements
 * **collector** UDP connections are now suffixed with `-udp` in
   destination target
+* **router** `send statistics to` construct was added to direct internal
+  statistics to a specific cluster
 
 ### Bugfixes
 * [Issue #159](https://github.com/grobian/carbon-c-relay/issues/159)